Household/Individual change history
