+++ src/country_workspace/workspaces/theme/static_src/src/styles.scss
@@ -5,6 +5,7 @@
 @import "./_footer.scss";
 @import "./_tabs.scss";
 @import "./_sidebar.scss";
+@import "./_history.scss";
 
 @tailwind base;
 @tailwind components;
