+++ tests/admin/test_admin_smoke.py
@@ -37,6 +37,7 @@ GLOBAL_EXCLUDED_MODELS = RegexList(
         "authtoken",
         "social_django",
         "depot",
+        "pghistory",
     ]
 )
